# apps/api/openapi/openapi-v1.json
{
...
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
    "/api/v1/aircraft" : {
      "get" : {
        "operationId" : "snapshot",
        "parameters" : [ {
          "in" : "query",
          "name" : "bbox",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "detail",
          "required" : false,
          "schema" : {
            "default" : "lite",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/geo+json" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "aircraft-controller" ]
      }
    },
    "/api/v1/aircraft/search" : {
      "get" : {
        "operationId" : "search",
        "parameters" : [ {
          "in" : "query",
          "name" : "q",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "aircraft-controller" ]
      }
    },
    "/api/v1/aircraft/{hex}" : {
      "get" : {
        "operationId" : "detail",
        "parameters" : [ {
          "in" : "path",
          "name" : "hex",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "aircraft-controller" ]
      }
    },
    "/api/v1/aircraft/{hex}/track" : {
      "get" : {
        "operationId" : "track",
        "parameters" : [ {
          "in" : "path",
          "name" : "hex",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date-time",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date-time",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "stepS",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/geo+json" : {
                "schema" : {
                  "additionalProperties" : { },
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "aircraft-controller" ]
      }
    },
...
  }
}